+++ src/main/java/tkcy/simpleaddon/loaders/recipe/handlers/harderstuff/Hydrogenation.java
@@ -3,18 +3,19 @@ package tkcy.simpleaddon.loaders.recipe.handlers.harderstuff;
 import static gregtech.api.GTValues.*;
 import static gregtech.api.unification.material.Materials.*;
 import static tkcy.simpleaddon.api.TKCYSAValues.SECOND;
+import static tkcy.simpleaddon.modules.PetroChemModule.desulfurizedFuels;
+import static tkcy.simpleaddon.modules.PetroChemModule.sulfuricLayers;
 
 import java.util.*;
 
 import net.minecraftforge.fluids.FluidStack;
 
-import gregtech.api.recipes.Recipe;
-import gregtech.api.recipes.RecipeMap;
-import gregtech.api.recipes.RecipeMaps;
+import gregtech.api.recipes.*;
 import gregtech.api.unification.material.Material;
 import gregtech.api.unification.ore.OrePrefix;
 
 import tkcy.simpleaddon.api.recipes.TKCYSARecipeMaps;
+import tkcy.simpleaddon.api.utils.CollectionHelper;
 import tkcy.simpleaddon.api.utils.RecipeHelper;
 
 public class Hydrogenation {
@@ -23,13 +24,9 @@ public class Hydrogenation {
         removeOilDesulfurizationRecipes();
         removeAmmoniaRecipe();
 
-        Map<Material, Material> sulfuricOilLayers = new HashMap<>();
-        sulfuricOilLayers.put(SulfuricHeavyFuel, HeavyFuel);
-        sulfuricOilLayers.put(SulfuricNaphtha, Naphtha);
-        sulfuricOilLayers.put(SulfuricLightFuel, LightFuel);
-        sulfuricOilLayers.put(SulfuricGas, RefineryGas);
+        CollectionHelper.buildMap(sulfuricLayers, desulfurizedFuels)
+                .forEach(Hydrogenation::oilDesulfurizationRecipes);
 
-        sulfuricOilLayers.forEach(Hydrogenation::oilDesulfurizationRecipes);
         miscHydrogenations();
     }